Fixed bug in OCR and created new dataset
Dataset will not be able through this repository anymore, since it is getting to big.
Fixed other smaller bugs and decided to make it able to use EMNIST and COMNIST data.
Now it is possible to save the model and import it when needed.
Validation set now is not a subset of train set but rather a disjunct set from train set.